Remove redundant static_cast in memcpy calls
Replace explicit static_cast<size_t>(...) with the original imageSize/bufferSize variables in memcpy calls across texture and pipeline code. Affected files: Texture2D.cpp, Texture3D.cpp, TextureCubemap.cpp, DotsPipeline.cpp. This removes unnecessary casts, clarifies intent, and avoids redundant conversions while preserving behavior.

+++ source/components/assets/textures/Texture2D.cpp
@@ -45,7 +45,7 @@ namespace vke {
                           stagingBuffer, stagingBufferMemory);
 
     Buffers::doMappedMemoryOperation(stagingBufferMemory, [pixels, imageSize](void* data) {
-      memcpy(data, pixels, static_cast<size_t>(imageSize));
+      memcpy(data, pixels, imageSize);
     });
 
     stbi_image_free(pixels);
